Move inner classes to upper level.
