+++ registration/src/test/java/org/spine3/samples/lobby/registration/seat/availability/TestSeatsAvailabilityAggregate.java
@@ -1,0 +1,78 @@
+/*
+ * Copyright 2015, TeamDev Ltd. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package org.spine3.samples.lobby.registration.seat.availability;
+
+import com.google.protobuf.Message;
+
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+
+import static com.google.common.base.Throwables.propagate;
+
+/**
+ * The aggregate used in {@link SeatsAvailabilityAggregate} tests.
+ *
+ * @author Alexander Litus
+ */
+@SuppressWarnings("TypeMayBeWeakened")
+public class TestSeatsAvailabilityAggregate extends SeatsAvailabilityAggregate {
+
+    public TestSeatsAvailabilityAggregate(SeatsAvailabilityId id) {
+        super(id);
+    }
+
+    // Is overridden to make it accessible in tests.
+    @Override
+    public void incrementState(SeatsAvailability newState) {
+        super.incrementState(newState);
+    }
+
+    public void apply(SeatsReserved event) {
+        invokeApplyMethod(event);
+    }
+
+    public void apply(SeatsReservationCommitted event) {
+        invokeApplyMethod(event);
+    }
+
+    public void apply(SeatsReservationCancelled event) {
+        invokeApplyMethod(event);
+    }
+
+    public void apply(AddedAvailableSeats event) {
+        invokeApplyMethod(event);
+    }
+
+    public void apply(RemovedAvailableSeats event) {
+        invokeApplyMethod(event);
+    }
+
+    private void invokeApplyMethod(Message event) {
+        try {
+            //noinspection DuplicateStringLiteralInspection
+            final Method apply = SeatsAvailabilityAggregate.class.getDeclaredMethod("apply", event.getClass());
+            apply.setAccessible(true);
+            apply.invoke(this, event);
+        } catch (NoSuchMethodException | InvocationTargetException | IllegalAccessException e) {
+            propagate(e);
+        }
+    }
+}
